(refactor) place _redis in caching folder
